Moved models to a sepparate folder
