+++ University_Simulator/src/main/java/com/challenge/alkemy/repositories/UserRepository.java
@@ -1,6 +1,6 @@
 package com.challenge.alkemy.repositories;
 
-import com.challenge.alkemy.User;
+import com.challenge.alkemy.models.User;
 
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
