+++ University_Simulator/src/main/java/com/challenge/alkemy/temp/Admin.java
@@ -1,6 +1,6 @@
 package com.challenge.alkemy.temp;
 
-import com.challenge.alkemy.User;
+import com.challenge.alkemy.models.User;
 
 public class Admin extends User {
